[ALELO] Melhorias no tipo utilizado (long -> bigdecimal - mais precisão e facilidade no cálculo)

+++ src/main/java/br/com/alelo/consumer/consumerpat/converter/card/CardSpendingConverter.java
@@ -1,5 +1,7 @@
 package br.com.alelo.consumer.consumerpat.converter.card;
 
+import java.math.BigDecimal;
+
 import br.com.alelo.consumer.consumerpat.dto.card.CardDebitBalanceRequestDTO;
 import br.com.alelo.consumer.consumerpat.dto.card.CardDebitBalanceResponseDTO;
 import br.com.alelo.consumer.consumerpat.model.card.PersistentCardSpending;
@@ -8,7 +10,7 @@ import br.com.alelo.consumer.consumerpat.model.card.PersistentConsumerCard;
 public interface CardSpendingConverter
 {
     PersistentCardSpending toModel(
-        Long totalDebit,
+        BigDecimal totalDebit,
         PersistentConsumerCard consumerCard,
         CardDebitBalanceRequestDTO cardDebitBalanceRequestDTO );
