+++ src/main/java/br/com/alelo/consumer/consumerpat/dto/card/ConsumerCardDTO.java
@@ -1,9 +1,11 @@
 package br.com.alelo.consumer.consumerpat.dto.card;
 
+import java.math.BigDecimal;
+
 public record ConsumerCardDTO(
     Integer id,
     Long number,
-    Long balanceValueCents,
+    BigDecimal balanceValueCents,
     String cardEstablishmentType )
 {
 }
